Yolov7-Dronedetection-Docker/requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2 # 3.4.3
numpy>=1.18.5,<1.24.0 # 1.21.2
opencv-python>=4.1.1 # 4.8.0.76
Pillow>=7.1.2 # 8.2.0
PyYAML>=5.3.1 # 5.4.1
requests>=2.23.0 # 2.26.0
scipy>=1.4.1 # 1.6.3
torch>=1.7.0,!=1.12.0 # 1.10.0a0+0aef44c
torchvision>=0.8.1,!=0.13.0 # 0.11.0a0
tqdm>=4.41.0 # 4.62.3
protobuf<4.21.3 # 3.18.1

# Logging -------------------------------------
tensorboard>=2.4.1 # 2.6.0
# wandb

# Plotting ------------------------------------
pandas>=1.1.4 # 2.0.3
seaborn>=0.11.0 # 0.12.2

# NiceGui -------------------------------------
Markdown==3.1.1
nicegui

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow
